requirements.txt
streamlit==1.28.1
streamlit-authenticator==0.2.3
stripe==7.0.0
altair==5.1.2
pandas==2.1.3
yfinance==0.2.32
pyyaml==6.0.1
requests==2.31.0
bcrypt==4.1.1
python-dotenv==1.0.0
plotly==5.18.0
sqlalchemy==2.0.23
